backend/requirements.txt
pyautogen==0.7.4
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.4.2
autogen-agentchat
autogen-ext[openai,azure]